+++ atlas-institute/src/interfaces.js
@@ -1,0 +1,53 @@
+export class ExecutiveKnowledgeInterface {
+  getHistory() {
+    throw new Error('Not implemented');
+  }
+}
+
+export class ResearchKnowledgeInterface {
+  retrieveResearchSummaries(_query) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class MemoryKnowledgeInterface {
+  retrieve(_query) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class MetricsKnowledgeInterface {
+  retrieveMetrics(_query) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class PerformanceKnowledgeInterface {
+  retrieveIntelligence(_query) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class ApprovalKnowledgeInterface {
+  getApprovalHistory(_query) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class CapabilityRegistryKnowledgeInterface {
+  listCapabilities() {
+    throw new Error('Not implemented');
+  }
+}
+
+export class WorkerOrchestrationKnowledgeInterface {
+  getHistory() {
+    throw new Error('Not implemented');
+  }
+}
+
+export class ControlCenterKnowledgeInterface {
+  getHistory() {
+    throw new Error('Not implemented');
+  }
+}